// src/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IAboutMeFields {
  /** Copy */
  copy: Document;
}

export interface IAboutMe extends Entry<IAboutMeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'aboutMe';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHeroFields {
  /** Heading */
  heading: Document;

  /** Sub-heading */
  subHeading: Document;
}

export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'hero';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IPortfolioItemFields {
  /** Title */
  title: string;

  /** Client */
  client: string;

  /** URL */
  url?: string | undefined;

  /** Github URL */
  githubUrl?: string | undefined;

  /** Copy */
  copy?: Document | undefined;

  /** Made with */
  madeWith?:
    | (
        | '(S)CSS Modules'
        | 'CSS'
        | 'Express.js'
        | 'Handlebars'
        | 'HTML'
        | 'JavaScript'
        | 'JavaScript (ES6)'
        | 'jQuery'
        | 'MongoDB'
        | 'Next.js'
        | 'Node.js'
        | 'React'
        | 'REST API (Next.js)'
        | 'Sass'
        | 'Styled Components'
        | 'TypeScript'
        | 'Webpack'
      )[]
    | undefined;

  /** Tested with */
  testedWith?:
    | (
        | 'Aquasec'
        | 'Chai/Mocha'
        | 'Cypress'
        | 'Enzyme'
        | 'Jest'
        | 'Nightwatch'
        | 'NVDA'
        | 'Postman'
        | 'Rapid7'
        | 'React Testing Library'
        | 'Veracode'
      )[]
    | undefined;

  /** Built with */
  builtWith?:
    | (
        | '.NET'
        | 'AWS (CI/CD)'
        | 'Contentful'
        | 'Github'
        | 'Netlify'
        | 'Sharepoint'
        | 'Swagger'
        | 'Vercel'
      )[]
    | undefined;

  /** Date */
  date: string;

  /** Is archived? */
  isArchived?: boolean | undefined;

  /** Omit */
  omit?: boolean | undefined;

  /** Position */
  position?: number | undefined;

  /** Image */
  image?: Asset | undefined;
}

/** Portfolio Item */

export interface IPortfolioItem extends Entry<IPortfolioItemFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'portfolioItem';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'aboutMe' | 'hero' | 'portfolioItem';

export type IEntry = IAboutMe | IHero | IPortfolioItem;

export type LOCALE_CODE = 'en-GB';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-GB';